Add a Resources section to the documentation (#23)
Adds a section 'Resources' with two pages 'Dissect in Action' and 'Talks and Conferences'. The first links to external websites that blog about Dissect while the latter links to past and upcoming talks and conferences we give on Dissect.

In addition, it is now possible to (re)build the core documentation without having to (re)build the API references of all the submodules which saves a tremendous amount of building time when working on the documentation.